+++ src/Http/ResponseDecoderInterface.php
@@ -8,4 +8,6 @@ use Psr\Http\Message\ResponseInterface;
 interface ResponseDecoderInterface
 {
     public function decode(ResponseInterface $httpResponse, string $responseClass = Response::class): Response;
+
+    public function omg();
 }