+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 
 ## Unreleased
+
+## 1.6.0 - 2018-06-01
 ### Added
 - `UserForget` command to anonymize or completely delete all data belonging to specific user(s) (`$matej->request()->forget()`).